+++ src/helpers/CustomObjectHelper.ts
@@ -1,0 +1,11 @@
+/**
+ * @file CustomObjectHelper.ts
+ * @fileoverview - Exports a single class that can be used to help with OCAPI
+ *    call operations for managing custom object definitions.
+ */
+
+export default class CustomObjectHelper {
+  constructor(args) {
+    /** @todo */
+  }
+}